Update containers.go
Added a new label "tsdproxy.container_host". This should be the same as the container name, and Docker DNS will resolve it. If this label is present then it overrides the value of container.TargetHostname. Doing this way makes sure that the dashboard works when this label is not present. Tested and works.